Add GitHub CI for tests
Initially I planned to do the releaes management via Actions and GPR.
Unfortunately GPR creates tags and releases, which makes it hard to
create a normal CI/CD workflow, like automatically building a docker
image for each push and tagging images based on code tags.
The problem is, that after pushing a tagged image, a new tag would be
created by GPR which makes this workflow unfeasible.
Thus I decided to create images and tags manually to sidestep that
problem.